move libnavigation-core sources to the ".core" package
