+++ examples/src/main/java/com/mapbox/navigation/examples/activity/TripSessionActivityKt.kt
@@ -12,17 +12,17 @@ import com.mapbox.mapboxsdk.location.modes.RenderMode
 import com.mapbox.mapboxsdk.maps.MapboxMap
 import com.mapbox.mapboxsdk.maps.OnMapReadyCallback
 import com.mapbox.mapboxsdk.maps.Style
-import com.mapbox.navigation.MapboxDistanceFormatter
 import com.mapbox.navigation.base.options.NavigationOptions
 import com.mapbox.navigation.base.typedef.METRIC
 import com.mapbox.navigation.base.typedef.ROUNDING_INCREMENT_FIFTY
 import com.mapbox.navigation.base.typedef.TWENTY_FOUR_HOURS
+import com.mapbox.navigation.core.MapboxDistanceFormatter
+import com.mapbox.navigation.core.trip.service.MapboxTripService
+import com.mapbox.navigation.core.trip.session.LocationObserver
+import com.mapbox.navigation.core.trip.session.MapboxTripSession
 import com.mapbox.navigation.examples.R
 import com.mapbox.navigation.navigator.MapboxNativeNavigatorImpl
 import com.mapbox.navigation.trip.notification.MapboxTripNotification
-import com.mapbox.navigation.trip.service.MapboxTripService
-import com.mapbox.navigation.trip.session.LocationObserver
-import com.mapbox.navigation.trip.session.MapboxTripSession
 import kotlinx.android.synthetic.main.activity_trip_session.*
 import timber.log.Timber
