+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/directions/session/RouteObserver.kt
@@ -1,0 +1,11 @@
+package com.mapbox.navigation.core.directions.session
+
+import com.mapbox.api.directions.v5.models.DirectionsRoute
+
+interface RouteObserver {
+    fun onRoutesChanged(routes: List<DirectionsRoute>)
+
+    fun onRoutesRequested()
+
+    fun onRoutesRequestFailure(throwable: Throwable)
+}
