+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/session/LocationObserver.kt
@@ -1,0 +1,8 @@
+package com.mapbox.navigation.core.trip.session
+
+import android.location.Location
+
+interface LocationObserver {
+    fun onRawLocationChanged(rawLocation: Location)
+    fun onEnhancedLocationChanged(enhancedLocation: Location)
+}
