+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/session/OffRouteObserver.kt
@@ -1,0 +1,5 @@
+package com.mapbox.navigation.core.trip.session
+
+interface OffRouteObserver {
+    fun onOffRouteStateChanged(offRoute: Boolean)
+}
