+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/session/RouteProgressObserver.kt
@@ -1,0 +1,7 @@
+package com.mapbox.navigation.core.trip.session
+
+import com.mapbox.navigation.base.trip.model.RouteProgress
+
+interface RouteProgressObserver {
+    fun onRouteProgressChanged(routeProgress: RouteProgress)
+}
