+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/session/TripSessionStateObserver.kt
@@ -1,0 +1,6 @@
+package com.mapbox.navigation.core.trip.session
+
+interface TripSessionStateObserver {
+    fun onSessionStarted()
+    fun onSessionStopped()
+}
