+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/trip/session/VoiceInstructionsObserver.kt
@@ -1,0 +1,7 @@
+package com.mapbox.navigation.core.trip.session
+
+import com.mapbox.api.directions.v5.models.VoiceInstructions
+
+interface VoiceInstructionsObserver {
+    fun onNewVoiceInstructions(voiceInstructions: VoiceInstructions)
+}
