--- libnavigation-core/src/main/java/com/mapbox/navigation/trip/session/LocationObserver.kt
@@ -1,8 +1,0 @@
-package com.mapbox.navigation.trip.session
-
-import android.location.Location
-
-interface LocationObserver {
-    fun onRawLocationChanged(rawLocation: Location)
-    fun onEnhancedLocationChanged(enhancedLocation: Location)
-}
