--- libnavigation-core/src/main/java/com/mapbox/navigation/trip/session/OffRouteObserver.kt
@@ -1,5 +1,0 @@
-package com.mapbox.navigation.trip.session
-
-interface OffRouteObserver {
-    fun onOffRouteStateChanged(offRoute: Boolean)
-}
