--- libnavigation-core/src/main/java/com/mapbox/navigation/trip/session/RouteProgressObserver.kt
@@ -1,7 +1,0 @@
-package com.mapbox.navigation.trip.session
-
-import com.mapbox.navigation.base.trip.model.RouteProgress
-
-interface RouteProgressObserver {
-    fun onRouteProgressChanged(routeProgress: RouteProgress)
-}
