--- libnavigation-core/src/main/java/com/mapbox/navigation/trip/session/TripSession.kt
@@ -1,41 +1,0 @@
-package com.mapbox.navigation.trip.session
-
-import android.location.Location
-import com.mapbox.android.core.location.LocationEngine
-import com.mapbox.android.core.location.LocationEngineRequest
-import com.mapbox.api.directions.v5.models.DirectionsRoute
-import com.mapbox.navigation.base.trip.model.RouteProgress
-import com.mapbox.navigation.trip.service.TripService
-
-internal interface TripSession {
-
-    val tripService: TripService
-    val locationEngine: LocationEngine
-    val locationEngineRequest: LocationEngineRequest
-    var route: DirectionsRoute?
-
-    fun getRawLocation(): Location?
-    fun getEnhancedLocation(): Location?
-    fun getRouteProgress(): RouteProgress?
-
-    fun start()
-    fun stop()
-
-    fun registerLocationObserver(locationObserver: LocationObserver)
-    fun unregisterLocationObserver(locationObserver: LocationObserver)
-
-    fun registerRouteProgressObserver(routeProgressObserver: RouteProgressObserver)
-    fun unregisterRouteProgressObserver(routeProgressObserver: RouteProgressObserver)
-
-    fun registerOffRouteObserver(offRouteObserver: OffRouteObserver)
-    fun unregisterOffRouteObserver(offRouteObserver: OffRouteObserver)
-
-    fun registerStateObserver(stateObserver: TripSessionStateObserver)
-    fun unregisterStateObserver(stateObserver: TripSessionStateObserver)
-
-    fun registerBannerInstructionsObserver(bannerInstructionsObserver: BannerInstructionsObserver)
-    fun unregisterBannerInstructionsObserver(bannerInstructionsObserver: BannerInstructionsObserver)
-
-    fun registerVoiceInstructionsObserver(voiceInstructionsObserver: VoiceInstructionsObserver)
-    fun unregisterVoiceInstructionsObserver(voiceInstructionsObserver: VoiceInstructionsObserver)
-}
