--- libnavigation-core/src/main/java/com/mapbox/navigation/trip/session/TripSessionStateObserver.kt
@@ -1,6 +1,0 @@
-package com.mapbox.navigation.trip.session
-
-interface TripSessionStateObserver {
-    fun onSessionStarted()
-    fun onSessionStopped()
-}
