--- libnavigation-core/src/main/java/com/mapbox/navigation/trip/session/VoiceInstructionsObserver.kt
@@ -1,7 +1,0 @@
-package com.mapbox.navigation.trip.session
-
-import com.mapbox.api.directions.v5.models.VoiceInstructions
-
-interface VoiceInstructionsObserver {
-    fun onNewVoiceInstructions(voiceInstructions: VoiceInstructions)
-}
